+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased changes (2026-08-17)
 
-* feat(smt): add yices solver plugin [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* fix(setups): stop prefixing z3's stdin with Z3_STATIC [[details](https://github.com/cbboyan/solverpy/commit/) | [browse](https://github.com/cbboyan/solverpy/tree/)]
+* feat(smt): add yices solver plugin [[details](https://github.com/cbboyan/solverpy/commit/eb0b477) | [browse](https://github.com/cbboyan/solverpy/tree/eb0b477)]
 * test(smt): cover the opensmt plugin [[details](https://github.com/cbboyan/solverpy/commit/1637c53) | [browse](https://github.com/cbboyan/solverpy/tree/1637c53)]
 * feat(smt): add opensmt solver plugin [[details](https://github.com/cbboyan/solverpy/commit/81c47a2) | [browse](https://github.com/cbboyan/solverpy/tree/81c47a2)]
 * feat(smt): add primo solver plugin [[details](https://github.com/cbboyan/solverpy/commit/e7ca1e8) | [browse](https://github.com/cbboyan/solverpy/tree/e7ca1e8)]